Correct class name of test cases (#15274)
