+++ shardingsphere-mode/shardingsphere-mode-core/src/test/java/org/apache/shardingsphere/mode/metadata/persist/service/impl/PropertiesPersistServiceTest.java
@@ -17,47 +17,31 @@
 
 package org.apache.shardingsphere.mode.metadata.persist.service.impl;
 
-import lombok.SneakyThrows;
-import org.apache.shardingsphere.infra.config.RuleConfiguration;
+import org.apache.shardingsphere.infra.config.props.ConfigurationPropertyKey;
 import org.apache.shardingsphere.mode.persist.PersistRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Collection;
-import java.util.stream.Collectors;
+import java.util.Properties;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public final class SchemaRuleMetaDataPersistServiceTest {
+public final class PropertiesPersistServiceTest {
+    
+    private static final String PROPS_YAML = ConfigurationPropertyKey.SQL_SHOW.getKey() + ": false\n";
     
     @Mock
     private PersistRepository repository;
     
     @Test
-    public void assertLoadWithoutExistedNode() {
-        assertTrue(new SchemaRulePersistService(repository).load("foo_db").isEmpty());
-    }
-    
-    @Test
-    public void assertLoadWithExistedNode() {
-        when(repository.get("/metadata/foo_db/rules")).thenReturn(readYAML());
-        Collection<RuleConfiguration> actual = new SchemaRulePersistService(repository).load("foo_db");
-        assertThat(actual.size(), is(1));
-    }
-    
-    @SneakyThrows({IOException.class, URISyntaxException.class})
-    private String readYAML() {
-        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource("yaml/persist/data-schema-rule.yaml").toURI()))
-                .stream().map(each -> each + System.lineSeparator()).collect(Collectors.joining());
+    public void assertLoad() {
+        when(repository.get("/props")).thenReturn(PROPS_YAML);
+        Properties actual = new PropertiesPersistService(repository).load();
+        assertThat(actual.get(ConfigurationPropertyKey.SQL_SHOW.getKey()), is(Boolean.FALSE));
     }
 }
